+++ src/graphql/mutations/local/vote.js
@@ -2,6 +2,6 @@ import gql from 'graphql-tag';
 
 export default gql`
   mutation voteLocal($procedureId: String!, $selection: VoteSelection!) {
-    votesLocal(procedureId: $procedureId, selection: $selection) @client
+    voteLocal(procedureId: $procedureId, selection: $selection) @client
   }
 `;