+++ src/graphql/queries/votedLocal.js
@@ -1,9 +1,0 @@
-import gql from 'graphql-tag';
-
-export default gql`
-  query votedLocal($procedureId: String) {
-    votedLocal(procedureId: $procedureId) @client {
-      selection
-    }
-  }
-`;