+++ src/migrations/index.js
@@ -1,3 +1,0 @@
-import presistProcedures from './presistProcedures';
-
-export default () => Promise.all([presistProcedures()]);