+++ src/screens/VoteVarification/BallotBox.js
@@ -8,10 +8,10 @@ import { Navigator, Navigation } from 'react-native-navigation';
 import VoteButton from '../../components/VoteButton';
 
 import VOTE from '../../graphql/mutations/vote';
-import VOTE_LOCAL from '../../graphql/mutations/voteLocal';
+import VOTE_LOCAL from '../../graphql/mutations/local/vote';
 import VOTED from '../../graphql/queries/voted';
 import VOTES from '../../graphql/queries/votes';
-import VOTED_LOCAL from '../../graphql/queries/votedLocal';
+import VOTE_SELECTION_LOCAL from '../../graphql/queries/local/voteSelection';
 import F_ACTIVITY_INDEX from '../../graphql/fragments/ProcedureActivityIndex';
 import F_VOTED from '../../graphql/fragments/ProcedureVoted';
 import VOTES_LOCAL from '../../graphql/queries/votesLocalKeyStore';
@@ -274,7 +274,7 @@ export default compose(
             variables: { procedureId, selection },
             refetchQueries: [
               {
-                query: VOTED_LOCAL,
+                query: VOTE_SELECTION_LOCAL,
                 variables: { procedureId },
               },
               {